requirements.txt
streamlit==1.12.0
pyarrow
snowflake-connector-python
pandas
cryptography